backend/requirements.txt
fastapi
uvicorn
pdfplumber
scikit-learn
python-dotenv
requests
python-multipart